feat : 화면 좌우 클릭으로 이동방식 변경 , 보스 흡혈, Hp바 색, 액티브 UI 추가
캐릭터 이동방식 변경
흡혈 스킬 구현
체력 비율에 따른 hp바 색 변화
유닛 액티브 UI 만들기(틀만 만들었음)

+++ PalaDog_Project/Assets/02.Scripts/02.Survival/Structs.cs
@@ -26,7 +26,7 @@ public struct SkillEffectEntry
 public struct SkillEffect
 {
     public int index;
-    public int value;
+    public float value;
     public float duration;
 }
 
@@ -37,7 +37,7 @@ public struct SkillEntry
     public int index;
     public int group;
     public SkillAct act;
-    public int coolTime;
+    public float coolTime;
     public TargetSearchType target_search_type;
     public UnitType target_type;
     public int target_search_num;
@@ -74,9 +74,9 @@ public struct CommonStatus
 {
     public int index;
     public string name;
-    public int HP;
+    public float HP;
     public float moveSpeed;
-    public int atk;
+    public float atk;
     public float atkRange;
     public float atkSpeed;
     public int[] skill;
